% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_module_score}
\alias{compute_module_score}
\title{Tailored module score calculation}
\usage{
compute_module_score(seu, features, name, seed = 1, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{features}{Vector of features (e.g. genes) to compute module score.}

\item{name}{Name of module score to be stored in metadata slot.}

\item{seed}{Set a random seed, for reproducibility.}

\item{...}{Additional parameters passed to Seurat's AddModuleScore.}
}
\value{
A ggplot2 object.
}
\description{
This function adapts the AddModuleScore Seurat function to
compute module scores based on a group of genes of interest.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
