% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_seurat_object}
\alias{create_seurat_object}
\title{Create Seurat object based on sample metadata.}
\usage{
create_seurat_object(
  data_dir,
  sample_meta,
  sample_meta_filename = NULL,
  meta_colnames = c("donor", "condition", "pass_qc"),
  plot_dir = NULL,
  use_scrublet = TRUE,
  use_soupx = FALSE,
  tenx_dir = "/premrna_outs/",
  expected_doublet_rate = 0.06,
  min.cells = 10,
  min.features = 200,
  assay = "RNA",
  ...
)
}
\arguments{
\item{data_dir}{Parent directory where all sample 10x files are stores.
Think of it as project directory.}

\item{sample_meta}{Sample metadata information in a Data.frame like object.
Columns should at least contain 'sample', 'donor', 'condition' and 'pass_qc'.}

\item{sample_meta_filename}{Filename of sample metadata information, same as 'meta'
parameter above. User should provide one of 'meta' or 'meta_filename'.}

\item{meta_colnames}{Sample metadata column names to store in Seurat metadata.}

\item{plot_dir}{Directory for storing QC plots. Used if use_soupx = TRUE.}

\item{use_scrublet}{Logical, wether to use Scrublet for doublet detection.}

\item{use_soupx}{Logical, wether to use SoupX for ambient RNA removal.}

\item{tenx_dir}{Name of 10x directory that count matrices are stored.
Default 'outs' for whole-cell data and 'premrna_outs' for single-nuclei.
If NULL, it checks the 'technology' column in sample metadata to infer the
sequencing technology.}

\item{expected_doublet_rate}{The expected fraction of transcriptomes that
are doublets, typically 0.05 - 0.1}

\item{min.cells}{Include features/genes detected in at least this many cells.}

\item{min.features}{Include cells where at least this many features/genes
are detected.}

\item{assay}{Assay to perform analysis, default "RNA".}

\item{...}{Additional named parameters passed to Seurat, Scrublet or SoupX.}
}
\value{
List of Seurat objects equal the number of samples in
the sample metadata file. If a single sample, returns a Seurat object.
}
\description{
This function creates Seurat object for each sample
in the metadata file. It also allows the user to perform SoupX (ambient
RNA removal) and Scrublet (doublet detection) analysis.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
