% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{dim_plot}
\alias{dim_plot}
\title{Tailored dimensional reduction plot}
\usage{
dim_plot(
  seu,
  reduction = "umap",
  group.by = "active.ident",
  split.by = NULL,
  ncol = NULL,
  legend.position = "right",
  col_pal = NULL,
  dims_plot = c(1, 2),
  pt.size = 1.4,
  label = FALSE,
  label.size = 7,
  combine = TRUE,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{reduction}{Which dimensionality reduction to use (required).}

\item{group.by}{Name of metadata column to group (color) cells by (required).}

\item{split.by}{Name of a metadata column to split plot by.}

\item{ncol}{Number of columns for display when combining plots.}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{col_pal}{Discrete colour palette to use, default is Hue palette
(hue_pal) from 'scales' package. Should be of equal length to number of
groups in 'group.by'.}

\item{dims_plot}{Dimensions to plot, must be a two-length numeric vector
specifying x- and y-dimensions.}

\item{pt.size}{Adjust point size for plotting.}

\item{label}{Whether to label the clusters in 'plot_reduction' space.}

\item{label.size}{Sets size of labels.}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}

\item{...}{Additional parameters passed to Seurat's DimPlot.}
}
\value{
A ggplot2 object.
}
\description{
This function extends the DimPlot Seurat function by
providing additional plotting options.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
