% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{dot_plot}
\alias{dot_plot}
\title{Tailored dot plot}
\usage{
dot_plot(
  seu,
  features,
  group.by = NULL,
  labels = NULL,
  xlab = "Signature",
  ylab = "Cluster",
  legend.position = "right",
  col_pal = NULL,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{features}{Vector of features to plot.}

\item{group.by}{Name of metadata column to group cells by (required).}

\item{labels}{If we want to have different labels plotted for each feature.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{col_pal}{Continuous colour palette to use, default "RdYlBu".}

\item{...}{Additional parameters passed to Seurat's DotPlot.}
}
\value{
A ggplot2 object.
}
\description{
This function adapts the DotPlot Seurat function by
providing additional plotting options.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
