% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{feature_plot}
\alias{feature_plot}
\title{Tailored feature plot}
\usage{
feature_plot(
  seu,
  reduction = "umap",
  features = "nFeature_RNA",
  max.cutoff = "q98",
  ncol = 1,
  legend.position = "right",
  col_pal = NULL,
  dims_plot = c(1, 2),
  pt.size = 1.4,
  combine = TRUE,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{reduction}{Which dimensionality reduction to use (required).}

\item{features}{Vector of features to plot.}

\item{max.cutoff}{Vector of maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{ncol}{Number of columns for display when combining plots.}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{col_pal}{Continuous colour palette to use, default "RdYlBu".}

\item{dims_plot}{Dimensions to plot, must be a two-length numeric vector
specifying x- and y-dimensions.}

\item{pt.size}{Adjust point size for plotting.}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}

\item{...}{Additional parameters passed to Seurat's FeaturePlot.}
}
\value{
A ggplot2 object.
}
\description{
This function adapts the FeaturePlot Seurat function by
providing additional plotting options.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
