% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{feature_plot_tailored}
\alias{feature_plot_tailored}
\title{Tailored feature plot}
\usage{
feature_plot_tailored(
  seu,
  feature,
  max.cutoff = "q98",
  min.cutoff = NA,
  reduction = "umap",
  slot = "data",
  col_pal = NULL,
  pt.size = 2,
  stroke = 0.05,
  legend.position = "right"
)
}
\arguments{
\item{seu}{Seurat object}

\item{feature}{Feature to plot.}

\item{max.cutoff}{Maximum cutoff value for feature, may specify quantile
in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{min.cutoff}{Minimum cutoff value for feature, may specify quantile
in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{reduction}{Dimensionality reduction to use.}

\item{slot}{Slot to extract data from.}

\item{col_pal}{Continuous colour palette to use, default "RdYlBu".}

\item{pt.size}{Adjust point size for plotting.}

\item{stroke}{Stroke value for each point.}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}
}
\value{
A ggplot2 object.
}
\description{
This function generates the same plot as \code{feature_plot},
although it focuses on a single feature and generates slightly better
looking plot.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
