% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_all_markers}
\alias{find_all_markers}
\title{Wrapper FindAllMarkers function}
\usage{
find_all_markers(seu, random.seed = 1, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{random.seed}{Seed to use for reproducibility purposes.}

\item{...}{Additional named parameters to Seurat FindAllMarkers function.}
}
\value{
Data frame with marker genes for each cluster and associated
statistics.
}
\description{
This function just calls the Seurat FindAllMarkers function.
Used only for naming consistency in this package.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
