% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_clusters}
\alias{find_clusters}
\title{Wrapper FindClusters function}
\usage{
find_clusters(seu, resolution, random.seed = 1, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{resolution}{Value of the resolution parameter, use a value
above (below) 1.0 if you want to obtain a larger (smaller) number of
communities.}

\item{random.seed}{Seed to use for reproducibility purposes.}

\item{...}{Additional named parameters to Seurat FindClusters function.}
}
\value{
Seurat object.
}
\description{
This function just calls the Seurat FindClusters function.
Used only for naming consistency in this package.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
