% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Tailored function for finding neighbors in lower dimensional space.}
\usage{
find_neighbors(seu, dims, reduction, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{dims}{Vector denoting which dimensions to use as input
features, e.g. dims = 1:30}

\item{reduction}{Reduction object to run UMAP. E.g. can be 'pca'
or 'harmony' for integrated data (required).}

\item{...}{Additional named parameters to Seurat FindNeighbors function.}
}
\value{
Seurat object.
}
\description{
This function adapts the Seurat FindNeighbors function by also
checking if 'dims' is larger than dimensions in 'reduction' object.
If yes, it takes maximum columns of 'reduction' object.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
