% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Tailored heatmap plot}
\usage{
heatmap_plot(seu, markers, topn = 10, assay = "RNA", filename = NULL, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{markers}{Data frame with marker genes for each cluster.
Expects a format as the output of Seurat's FindAllMarkers.}

\item{topn}{Top N marker genes to plot for each cluster.}

\item{assay}{Assay to extract gene expression data from (default 'RNA').}

\item{filename}{Filename for saving the heatmap plot. If null, the}

\item{...}{Additional parameters passed to 'pheatmap' function}
}
\value{
A ggplot2 object.
}
\description{
Hetmap plot showing marker genes after the clustering process.
Expression levels are taken from the \code{scaled.data} slot.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
