% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lognormalize_and_pca}
\alias{lognormalize_and_pca}
\alias{lognormalize_and_pca.list}
\alias{lognormalize_and_pca.Seurat}
\title{Log normalisation and PCA computation}
\usage{
lognormalize_and_pca(seu, ...)

\method{lognormalize_and_pca}{list}(seu, npcs = NULL, n_hvgs = 3000, assay = "RNA", ...)

\method{lognormalize_and_pca}{Seurat}(seu, npcs = NULL, n_hvgs = 3000, assay = "RNA", ...)
}
\arguments{
\item{seu}{Seurat object or list of Seurat objects(required).}

\item{...}{Additional named parameters passed to Seurat functions.}

\item{npcs}{Number of Principal Components (PCs) to compute.
If NULL, PCA analysis is skipped.}

\item{n_hvgs}{Number of highly variable genes (HVGs) to compute. Only HVGs
will be used as input to PCA.}

\item{assay}{Assay to perform analysis, default "RNA".}
}
\value{
List of processed Seurat objects. If a single sample, returns
a processed Seurat object instead of a list.
}
\description{
This function wrapps the default Seurat analysis of 1.
data normalisation, 2. identification of HVGs, 3. scaling expression
and computing PCA. If a list of Seurat objects is provided
this is performed for each element in the list.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
