% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{pca_feature_cor_plot}
\alias{pca_feature_cor_plot}
\title{PCA and feature metadata correlation heatmap plot}
\usage{
pca_feature_cor_plot(seu, features)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{features}{Vector of features in metadata to plot.}
}
\value{
A ggplot2 object.
}
\description{
This function computes correlation of principal components
with specific feature metadata (e.g. nFeature) to identify potential
technical/batch effects in the data.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
