% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qc_filter_seurat_object}
\alias{qc_filter_seurat_object}
\alias{qc_filter_seurat_object.list}
\alias{qc_filter_seurat_object.Seurat}
\title{Filter Seurat object based on QC metrics.}
\usage{
qc_filter_seurat_object(seu, nfeat_thresh, mito_thresh, ...)

\method{qc_filter_seurat_object}{list}(seu, nfeat_thresh, mito_thresh, ...)

\method{qc_filter_seurat_object}{Seurat}(seu, nfeat_thresh, mito_thresh, ...)
}
\arguments{
\item{seu}{Seurat object or list of Seurat objects(required).}

\item{nfeat_thresh}{Filter cells that have less than 'nfeat_thresh' counts
expressed.}

\item{mito_thresh}{Filter cells with more than 'mito_thresh'\% counts.}

\item{...}{Additional parameters.}
}
\value{
List of QC filtered Seurat objects. If a single sample, returns
a QC filtered Seurat object instead of a list.
}
\description{
This function filters a Seurat object (i.e. removes cells)
based on QC metric thresholds. For now only based on nFeatures and
mitochondrial percentage.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
