% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_harmony_pipeline}
\alias{run_harmony_pipeline}
\title{(Internal) pipeline for Harmony integration}
\usage{
run_harmony_pipeline(seu_obj, io, opts, ...)
}
\arguments{
\item{seu_obj}{Seurat object (required).}

\item{io}{Input/output settings defined in main function.}

\item{opts}{Options settings defined in main function.}

\item{...}{Additional named parameters passed to Seurat's or Harmony
functions.}
}
\value{
None.
}
\description{
This is an internal pipeline which wraps all analysis steps in
a function. This is meant to be used as a guide and in each analysis, users
should write their own function (possibly adapting the code of this
function). This is the reason this function is not explicitly 'exported',
i.e. even if you load the library, you cannot directly call the function,
but you should call seurat_pipeline:::run_harmony_pipeline. Elements in 'io'
and 'opts' are not checked, so function might return errors if not all
required values are not defined.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
