% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_qc_pipeline}
\alias{run_qc_pipeline}
\title{QC pipeline}
\usage{
run_qc_pipeline(
  data_dir,
  sample_meta,
  sample_meta_filename = NULL,
  nfeat_thresh = 500,
  mito_thresh = 5,
  meta_colnames = c("donor", "condition", "pass_qc"),
  results_dir = NULL,
  qc_to_plot = c("nFeature_RNA", "nCount_RNA", "percent.mito"),
  use_scrublet = TRUE,
  use_soupx = FALSE,
  tenx_dir = "/premrna_outs/",
  expected_doublet_rate = 0.06,
  force_reanalysis = TRUE,
  min.cells = 10,
  min.features = 200,
  assay = "RNA",
  ...
)
}
\arguments{
\item{data_dir}{Parent directory where all sample 10x files are stores.
Think of it as project directory.}

\item{sample_meta}{Sample metadata information in a Data.frame like object.
Columns should at least contain 'sample', 'donor', 'condition' and 'pass_qc'.}

\item{sample_meta_filename}{Filename of sample metadata information, same as 'meta'
parameter above. User should provide one of 'meta' or 'meta_filename'.}

\item{nfeat_thresh}{Filter cells that have less than 'nfeat_thresh' counts
expressed.}

\item{mito_thresh}{Filter cells with more than 'mito_thresh'\% counts.}

\item{meta_colnames}{Sample metadata column names to store in Seurat metadata.}

\item{results_dir}{Output directory for storing analysis results.}

\item{qc_to_plot}{Vector of features in metadata to plot.}

\item{use_scrublet}{Logical, wether to use Scrublet for doublet detection.}

\item{use_soupx}{Logical, wether to use SoupX for ambient RNA removal.}

\item{tenx_dir}{Name of 10x directory that count matrices are stored.
Default 'outs' for whole-cell data and 'premrna_outs' for single-nuclei.
If NULL, it checks the 'technology' column in sample metadata to infer the
sequencing technology.}

\item{expected_doublet_rate}{The expected fraction of transcriptomes that
are doublets, typically 0.05 - 0.1}

\item{force_reanalysis}{Logical, if intermediate file 'seu_preqc.rds' (with
created Seurat object) exists and force_reanalysis = FALSE, read object
instead of re-running whole analysis. Added for computing time efficiency
purposes.}

\item{min.cells}{Include features/genes detected in at least this many cells.}

\item{min.features}{Include cells where at least this many features/genes
are detected.}

\item{assay}{Assay to perform analysis, default "RNA".}

\item{...}{Additional named parameters passed to Seurat, Scrublet or SoupX.}
}
\value{
List of Seurat objects as the length of the number of samples in
the sample metadata file. If a single sample, return a Seurat object
instead of a list.
}
\description{
This function implements all the analysis steps for perfoming
QC. These include: 1. reading all sample information from metadata object/file
and generating one Seurat object per sample. 2. Performs SoupX (ambient
RNA removal) and Scrublet (doublet detection) if user defines the
corresponding parameters. 3. Filter Seurat object according to QC
criteria 4. Generate correspond QC plots.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
