% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{run_spatial_qc_pipeline}
\alias{run_spatial_qc_pipeline}
\title{Spatial QC pipeline}
\usage{
run_spatial_qc_pipeline(
  data_dir,
  sample_meta,
  sample_meta_filename = NULL,
  nfeat_thresh = 500,
  mito_thresh = 5,
  meta_colnames = c("donor", "condition", "pass_qc"),
  out_dir = NULL,
  qc_to_plot = c("nFeature_Spatial", "nCount_Spatial", "percent.mito"),
  alpha = c(0.1, 0.9),
  pt.size.factor = 1.1,
  max.cutoff = "q98",
  min.cutoff = NA,
  spatial_col_pal = "inferno",
  crop = FALSE,
  tenx_dir = "outs",
  obj_filename = "seu_qc",
  ...
)
}
\arguments{
\item{data_dir}{Parent directory where all sample 10x files are stored.
Think of it as project directory.}

\item{sample_meta}{Sample metadata information in a Data.frame like object.
Columns should at least contain 'sample', 'donor', 'condition' and 'pass_qc'.}

\item{sample_meta_filename}{Filename of sample metadata information, same as 'meta'
parameter above. User should provide one of 'meta' or 'meta_filename'.}

\item{nfeat_thresh}{Filter cells that have less than 'nfeat_thresh' counts
expressed.}

\item{mito_thresh}{Filter cells with more than 'mito_thresh'\% counts.}

\item{meta_colnames}{Sample metadata column names to store in Seurat metadata.}

\item{out_dir}{Output directory for storing analysis results.}

\item{qc_to_plot}{Vector of features in metadata to plot.}

\item{alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max range of values (between 0 and 1).}

\item{pt.size.factor}{Scale the size of the spots.}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may
specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may
specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{spatial_col_pal}{(Spatial) Continuous colour palette to use from viridis package to
colour spots on tissue, default "inferno".}

\item{crop}{(Spatial) Crop the plot in to focus on points plotted. Set to FALSE to
show entire background image.}

\item{tenx_dir}{Name of 10x base directory, e.g. with outputs after running
cellranger. Default 'premrna_outs', i.e. assumes single-nuclei RNA-seq.}

\item{obj_filename}{Filename of the stored Seurat object, default 'seu_qc'.}

\item{...}{Additional named parameters passed to Seurat, Scrublet or SoupX.}
}
\value{
List of Seurat objects as the length of the number of samples in
the sample metadata file. If a single sample, return a Seurat object
instead of a list.
}
\description{
This function implements all the analysis steps for perfoming
QC. These include: 1. reading all sample information from metadata object/file
and generating one Seurat object per sample. 2. Performs SoupX (ambient
RNA removal) and Scrublet (doublet detection) if user defines the
corresponding parameters. 3. Filter Seurat object according to QC
criteria 4. Generate correspond QC plots.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
