% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_umap}
\alias{run_umap}
\title{Tailored UMAP function}
\usage{
run_umap(seu, dims, reduction, seed = 1, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{dims}{Vector denoting which dimensions to use as input
features, e.g. dims = 1:30}

\item{reduction}{Reduction object to run UMAP. E.g. can be 'pca'
or 'harmony' for integrated data (required).}

\item{seed}{Set a random seed, for reproducibility.}

\item{...}{Additional named parameters passed to Seurat RunUMAP function.}
}
\value{
Seurat object.
}
\description{
This function adapts the Seurat RunUMAP function by also
checking if 'dims' is larger than dimensions in 'reduction' object.
If yes, it uses the maximum columns of 'reduction' object.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
