% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPipe_package.R
\docType{package}
\name{SeuratPipe}
\alias{SeuratPipe}
\title{\code{SeuratPipe}: Streamlining Seurat analysis}
\value{
SeuratPipe main package documentation.
}
\description{
SeuratPipe contains common analysis, processing and
plotting functions with an attempt to streamline single-cell genomics
analysis using the Seurat package.
}
\author{
C.A.Kapourani \email{kapouranis.andreas@gmail.com}
}
