% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{spatial_dim_plot}
\alias{spatial_dim_plot}
\title{Tailored spatial dim plot}
\usage{
spatial_dim_plot(
  seu,
  group.by = "active.ident",
  alpha = 0.6,
  pt.size.factor = 1.6,
  crop = TRUE,
  col_pal = NULL,
  legend.position = "top",
  combine = TRUE,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{group.by}{Name of meta.data column to group the data by.}

\item{alpha}{Controls opacity of spots. Provide a single alpha value for
each plot.}

\item{pt.size.factor}{Scale the size of the spots.}

\item{crop}{Crop the plot in to focus on points plotted. Set to FALSE to
show entire background image.}

\item{col_pal}{Discrete colour palette to use, default is Hue palette
(hue_pal) from 'scales' package. Should be of equal length to number of
groups in 'group.by'.}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}

\item{...}{Additional parameters passed to Seurat's SpatialDimPlot.}
}
\value{
A ggplot2 object.
}
\description{
This function adapts the SpatialDimPlot Seurat function by
providing additional plotting options.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
