% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{spatial_feature_plot}
\alias{spatial_feature_plot}
\title{Tailored spatial feature plot}
\usage{
spatial_feature_plot(
  seu,
  features = "nFeature_Spatial",
  title = NULL,
  alpha = c(0.1, 0.9),
  pt.size.factor = 1.1,
  ncol = NULL,
  max.cutoff = "q98",
  min.cutoff = NA,
  crop = FALSE,
  col_pal = "inferno",
  legend.position = "top",
  combine = TRUE,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{features}{Vector of features to plot.}

\item{title}{Plot title}

\item{alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max range of values (between 0 and 1).}

\item{pt.size.factor}{Scale the size of the spots.}

\item{ncol}{Number of columns for display when having multiple features.}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may
specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may
specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{crop}{Crop the plot in to focus on points plotted. Set to FALSE to
show entire background image.}

\item{col_pal}{Continuous colour palette to use from viridis package,
default "inferno".}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}

\item{...}{Additional parameters passed to Seurat's SpatialFeaturePlot.}
}
\value{
A ggplot2 object.
}
\description{
This function adapts the SpatialFeaturePlot Seurat function by
providing additional plotting options.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
