% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{subset_dim_plot}
\alias{subset_dim_plot}
\title{Subset dim plot}
\usage{
subset_dim_plot(
  seu,
  subset.by,
  reduction = "umap",
  ncol = NULL,
  col_pal = NULL,
  pt.size = 2,
  stroke = 0.05,
  back.pt.size = 0.5,
  back.alpha = 0.1,
  back.color = "grey",
  combine = TRUE
)
}
\arguments{
\item{seu}{Seurat object}

\item{subset.by}{Metadata column whose unique values will generate split
dim plots.}

\item{reduction}{Dimensionality reduction to use.}

\item{ncol}{Number of columns for display when having multiple features.}

\item{col_pal}{Discrete colour palette to use.}

\item{pt.size}{Adjust point size for plotting.}

\item{stroke}{Stroke value for each point.}

\item{back.pt.size}{Adjust background point size for plotting.}

\item{back.alpha}{Adjust opacity for background points.}

\item{back.color}{Colour for background points.}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}
}
\value{
A ggplot2 object.
}
\description{
This function splits the dim plots according to a grouping
variable defined in \code{subset.by}. The main difference is that the whole
dataset is also shown in the background as smaller points.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
