% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{cluster_analysis}
\alias{cluster_analysis}
\title{Common clustering analysis steps}
\usage{
cluster_analysis(
  seu,
  ndim = 20,
  res = seq(0.1, 0.1, by = 0.1),
  logfc.threshol = 0.5,
  min.pct = 0.25,
  only.pos = TRUE,
  topn = 10,
  plot_dir = NULL,
  plot_cluster_markers = TRUE,
  modules_group = NULL,
  cluster_reduction = "harmony",
  plot_reduction = "umap",
  max.cutoff = "q98",
  cl.legend.position = "right",
  legend.position = "top",
  col_pal = NULL,
  pt.size = 1.4,
  label = TRUE,
  label.size = 8,
  assay = "RNA",
  ctrl = 100,
  seed = 1,
  fig.res = 200,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{ndim}{Number of dimensions to use for clustering (from
cluster_reduction' object).}

\item{res}{Vector with clustering resolutions (e.g. seq(0.1, 0.6, by = 0.1)).}

\item{logfc.threshol}{Limit testing to genes which show, on average, at
least X-fold difference (log-scale) between the two groups of cells.}

\item{min.pct}{Only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations.}

\item{only.pos}{Only return positive markers (TRUE by default).}

\item{topn}{Top cluster marker genes to use for plot (in heatmap and
feature plots), default is 10.}

\item{plot_dir}{Directory to save generated plots. If NULL, plots are
not saved.}

\item{plot_cluster_markers}{Logical, wheather to create feature plots with
'topn' cluster markers. Added mostly to reduce number of files (and size)
in analysis folders. Default is TRUE.}

\item{modules_group}{Group of modules (named list of lists) storing features
(e.g. genes) to compute module score for each identified cluster. This step
can be useful for annotating the different clusters by saving dot
plots for each group. Assumes that we already have computed the modules e.g.
by calling the 'module_score_analysis' function. If 'plot_dir' is NULL,
no plots will be generated.}

\item{cluster_reduction}{Dimensionality reduction to use for performing
clustering. Default is 'harmony', should be set to 'pca' if we do not
perform data integration.}

\item{plot_reduction}{Dimensionality reduction to use for plotting
functions. Default is 'umap'.}

\item{max.cutoff}{Vector of maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{cl.legend.position}{Position of legend for clusters, default "right"
(set to "none" for clean plot)}

\item{legend.position}{Position of legend, default "top" (set to "none"
for clean plot).}

\item{col_pal}{Colour palette to use.}

\item{pt.size}{Adjust point size for plotting.}

\item{label}{Whether to label the clusters.}

\item{label.size}{Sets size of labels.}

\item{assay}{Assay to perform analysis, default "RNA".}

\item{ctrl}{Number of control features selected from the same bin per
analyzed feature.}

\item{seed}{Set a random seed.}

\item{fig.res}{Figure resolution in ppi (see 'png' function).}

\item{...}{Additional named parameters passed to Seurat
analysis and plotting functions, such as FindClusters, FindAllMarkers,
DimPlot and FeaturePlot.}
}
\value{
Updated Seurat object clustered cells
}
\description{
This function implements all the analysis steps for performing
clustering on a Seurat object. These include, 1. finding neighbours in lower
dimensional space (defined in 'cluster_reduction' parameter) 2. obtaining
clusters, 3. identifying marker genes (NOTE: to speed up re-analysis it
first checks if file with marker genes is already present, if yes reads the
file instead of calling FinaAllMarkers) and 4. generating plots, which
include heatmap with (scaled) expression of marker genes in each cluster,
marker gene expression on feature plots (e.g. UMAP space, defined in
plot_reduction'  parameter), dot / feature plots with
pre-computed module scores on each cluster (assumes we have first run
'module_score_analysis' function). This step could be useful for
lineage annotation.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
