% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Tailored heatmap plot}
\usage{
heatmap_plot(seu, markers, topn_genes = 10, filename = NULL, ...)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{markers}{Data frame with marker genes for each cluster.
Expects a format as the output of Seurat's FindAllMarkers.}

\item{topn_genes}{Top N marker genes to plot for each cluster.}

\item{filename}{Filename for saving the heatmap plot. If null, the heatmap
is just plotted in device.}

\item{...}{Additional parameters passed to 'pheatmap' function}
}
\value{
A ggplot2 object.
}
\description{
Heatmap plot showing marker genes after the clustering process.
Expression levels are taken from the \code{scaled.data} slot. Current
implementation is 'hard' coded: that is for heatmap column annotation,
it assumes the Seurat object contains 'condition', 'sample' and
'seurat_clusters' as columns in meta.data slot.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
