% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Tailored heatmap plot}
\usage{
heatmap_plot(
  seu,
  markers,
  topn_genes = 10,
  diff_cluster_pct = 0.1,
  pval_adj = 0.05,
  filename = NULL,
  col_pal = NULL,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{markers}{Data frame with marker genes for each cluster.
Expects a format as the output of Seurat's FindAllMarkers.}

\item{topn_genes}{Top N marker genes to plot for each cluster.}

\item{diff_cluster_pct}{Retain marker genes per cluster if their
\code{pct.1 - pct.2 > diff_cluster_pct}, i.e. they show cluster
specific expression. Set to -Inf, to ignore this additional filtering.}

\item{pval_adj}{Adjusted p-value threshold to consider marker genes per cluster.}

\item{filename}{Filename for saving the heatmap plot. If null, the heatmap
is just plotted in device.}

\item{col_pal}{Discrete colour palette to use, default is Hue palette
(hue_pal) from 'scales' package.}

\item{...}{Additional parameters passed to 'pheatmap' function}
}
\value{
A ggplot2 object.
}
\description{
Heatmap plot showing marker genes after the clustering process.
Expression levels are taken from the \code{scaled.data} slot. Current
implementation is 'hard' coded: that is for heatmap column annotation,
it assumes the Seurat object contains 'condition', 'sample' and
'seurat_clusters' as columns in meta.data slot.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
