% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{module_score_analysis}
\alias{module_score_analysis}
\title{Module score analysis}
\usage{
module_score_analysis(
  seu,
  modules_group,
  plot_dir = NULL,
  reduction = "umap",
  max.cutoff = "q98",
  legend.position = "top",
  col_pal = NULL,
  dims_plot = c(1, 2),
  seed = 1,
  ctrl = 100,
  pt.size = 1.4,
  fig.res = 200,
  ...
)
}
\arguments{
\item{seu}{Seurat object (required).}

\item{modules_group}{Group of modules (named list of lists) storing features
(e.g. genes) to compute module score for each identified cluster. This step
can be useful for annotating the different clusters by saving dot
plots for each group. Assumes that we already have computed the modules e.g.
by calling the 'module_score_analysis' function. If 'plot_dir' is NULL,
no plots will be generated.}

\item{plot_dir}{Directory to save generated plots. If NULL, plots are
not saved.}

\item{reduction}{Which dimensionality reduction to use (required).}

\item{max.cutoff}{Vector of maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile
(eg, 'q1', 'q10').}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{col_pal}{Continuous colour palette to use, default "RdYlBu".}

\item{dims_plot}{Dimensions to plot, must be a two-length numeric vector
specifying x- and y-dimensions.}

\item{seed}{Set a random seed, for reproducibility.}

\item{ctrl}{Number of control features selected from the same bin per
analyzed feature.}

\item{pt.size}{Adjust point size for plotting.}

\item{fig.res}{Figure resolution in ppi (see 'png' function).}

\item{...}{Additional named parameters passed to Seurat's AddModuleScore
and FeaturePlot}
}
\value{
Updated Seurat object with calculated module scores in metadata.
}
\description{
This function takes a group of modules and calculates module
scores for the expression programs. In addition, if \code{plot_dir} is not NULL
creates feature plots with marker genes and the module score and saved them
in \code{plot_dir} directory.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
