% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_cluster_pipeline}
\alias{run_cluster_pipeline}
\title{Pipeline for clustering analysis}
\usage{
run_cluster_pipeline(
  seu_obj,
  out_dir,
  npcs = c(50),
  ndims = c(30),
  res = seq(0.1, 0.3, by = 0.1),
  modules_group = NULL,
  metadata_to_plot = c("sample", "condition"),
  qc_to_plot = NULL,
  logfc.threshold = 0.5,
  min.pct = 0.25,
  only.pos = TRUE,
  topn_genes = 10,
  diff_cluster_pct = 0.1,
  pval_adj = 0.05,
  pcs_to_remove = NULL,
  plot_cluster_markers = TRUE,
  max.cutoff = "q98",
  min.cutoff = NA,
  n_hvgs = 3000,
  seed = 1,
  label = TRUE,
  label.size = 8,
  pt.size = 1.4,
  fig.res = 200,
  cont_col_pal = NULL,
  discrete_col_pal = NULL,
  cont_alpha = c(0.1, 0.9),
  discrete_alpha = 0.9,
  pt.size.factor = 1.1,
  spatial_col_pal = "inferno",
  crop = FALSE,
  plot_spatial_markers = FALSE,
  ...
)
}
\arguments{
\item{seu_obj}{Seurat object (required).}

\item{out_dir}{Output directory for storing analysis results.}

\item{npcs}{Number of principal components, can be a vector e.g. c(50, 70).}

\item{ndims}{Top PCA dimensions to perform UMAP and clustering,
can be a vector e.g. c(50, 70).}

\item{res}{Vector with clustering resolutions (e.g. seq(0.1, 0.6, by = 0.1)).}

\item{modules_group}{Group of modules (named list of lists) storing features
(e.g. genes) to compute module score for each identified cluster. This step
can be useful for annotating the different clusters by saving dot/feature
plots for each group.}

\item{metadata_to_plot}{Vector with metadata names to plot, they should be
present in the meta.data slot of the Seurat object.}

\item{qc_to_plot}{Vector with QC names to plot, they should be
present in the meta.data slot of the Seurat object.}

\item{logfc.threshold}{Limit testing to genes which show, on average, at
least X-fold difference (log-scale) between the two groups of cells.}

\item{min.pct}{Only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations.}

\item{only.pos}{Only return positive markers (TRUE by default).}

\item{topn_genes}{Top cluster marker genes to use for plotting (in heatmap and
feature plots), default is 10.}

\item{diff_cluster_pct}{Retain marker genes per cluster if their
\code{pct.1 - pct.2 > diff_cluster_pct}, i.e. they show cluster
specific expression. Set to -Inf, to ignore this additional filtering.}

\item{pval_adj}{Adjusted p-value threshold to consider marker genes per cluster.}

\item{pcs_to_remove}{Which PCs should be removed prior to performing clustering.
Possibly due to being correlated with technical/batch effects. If NULL,
all PCs are used.}

\item{plot_cluster_markers}{Logical, wheather to create feature plots with
'topn_genes' cluster markers. Added mostly to reduce number of files
(and size) in analysis folders. Default is TRUE.}

\item{max.cutoff}{Maximum cutoff values for plotting each continuous
feature, e.g. gene expression levels. May specify quantile in the form of
'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{min.cutoff}{Maximum cutoff values for plotting each continuous
feature, e.g. gene expression levels. May specify quantile in the form of
'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{n_hvgs}{Number of highly variable genes (HVGs) to compute, which will
be used as input to PCA.}

\item{seed}{Set a random seed, for reproducibility.}

\item{label}{Whether to label the clusters in 'plot_reduction' space.}

\item{label.size}{Sets size of labels.}

\item{pt.size}{Adjust point size for plotting.}

\item{fig.res}{Figure resolution in ppi (see 'png' function).}

\item{cont_col_pal}{Continuous colour palette to use, default "RdYlBu".}

\item{discrete_col_pal}{Discrete colour palette to use, default is Hue palette
(hue_pal) from 'scales' package.}

\item{cont_alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max range of values (between 0 and 1).}

\item{discrete_alpha}{Controls opacity of spots. Provide a single alpha value.}

\item{pt.size.factor}{Scale the size of the spots.}

\item{spatial_col_pal}{Continuous colour palette to use from viridis package to
colour spots on tissue, default "inferno".}

\item{crop}{Crop the plot in to focus on spots that passed QC plotted. Set to FALSE to
show entire background image.}

\item{plot_spatial_markers}{Logical, whether to create spatial feature plots
with expression of individual genes.}

\item{...}{Additional named parameters passed to Seurat functions.}
}
\value{
An updated Seurat object. Note that if multiple \code{npcs} and \code{ndims}
are given, only the last setting will be returned. All analysis results are
also stored on disk.
}
\description{
This function wraps the most common Seurat analysis pipeline
for cell type identification. These include: 1. data processing, e.g.
normalisation. 2. Running PCA. 3. Perform UMAP and clustering. Analysis
outputs are stored in corresponding directories. Note this pipeline
requires a single Seurat object/sample.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
