% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_harmony_pipeline}
\alias{run_harmony_pipeline}
\title{Pipeline for Harmony integration}
\usage{
run_harmony_pipeline(
  seu_obj,
  out_dir,
  batch_id = "sample",
  npcs = c(50),
  ndims = c(30),
  res = seq(0.1, 0.3, by = 0.1),
  modules_group = NULL,
  metadata_to_plot = c("sample", "condition"),
  qc_to_plot = NULL,
  logfc.threshold = 0.5,
  min.pct = 0.25,
  only.pos = TRUE,
  topn_genes = 10,
  pcs_to_remove = NULL,
  obj_filename = "seu_harmony",
  force_reanalysis = TRUE,
  plot_cluster_markers = TRUE,
  max.cutoff = "q98",
  n_hvgs = 3000,
  max.iter.harmony = 50,
  seed = 1,
  label = TRUE,
  label.size = 8,
  pt.size = 1.4,
  fig.res = 200,
  cont_alpha = c(0.1, 0.9),
  discrete_alpha = 0.6,
  pt.size.factor = 1.6,
  spatial_col_pal = "inferno",
  crop = TRUE,
  ...
)
}
\arguments{
\item{seu_obj}{Seurat object or list of Seurat objects(required).}

\item{out_dir}{Output directory for storing analysis results.}

\item{batch_id}{Name of batch to try and remove with data integration
(required). Can also be a vector if multiple batch information are present.
Should be a column name in Seurat 'meta.data'. Default is "sample".
This parameter is called 'group.by.vars' in Harmony.}

\item{npcs}{Number of principal components, can be a vector e.g. c(50, 70).}

\item{ndims}{Top Harmony dimensions to perform UMAP and clustering,
can be a vector e.g. c(50, 70).}

\item{res}{Vector with clustering resolutions (e.g. seq(0.1, 0.6, by = 0.1)).}

\item{modules_group}{Group of modules (named list of lists) storing features
(e.g. genes) to compute module score for each identified cluster. This step
can be useful for annotating the different clusters by saving dot/feature
plots for each group.}

\item{metadata_to_plot}{Vector with metadata names to plot, they should be
present in the meta.data slot of the Seurat object.}

\item{qc_to_plot}{Vector with QC names to plot, they should be
present in the meta.data slot of the Seurat object.}

\item{logfc.threshold}{Limit testing to genes which show, on average, at
least X-fold difference (log-scale) between the two groups of cells.}

\item{min.pct}{Only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations.}

\item{only.pos}{Only return positive markers (TRUE by default).}

\item{topn_genes}{Top cluster marker genes to use for plotting (in heatmap and
feature plots), default is 10.}

\item{pcs_to_remove}{Which PCs should be removed prior to running Harmony.
Possibly due to being correlated with technical/batch effects. If NULL,
all PCs are used.}

\item{obj_filename}{Filename of the stored Seurat object, default
'seu_harmony'. Number of PCs will be added to the filename automatically.}

\item{force_reanalysis}{Logical, if intermediate object 'seu_harmony_<>.rds'
exists and force_reanalysis = FALSE, read object instead of re-running
Harmony integration. Added for computing time efficiency purposes.}

\item{plot_cluster_markers}{Logical, wheather to create feature plots with
'topn_genes' cluster markers. Added mostly to reduce number of files
(and size) in analysis folders. Default is TRUE.}

\item{max.cutoff}{Maximum cutoff values for plotting each continuous
feature, e.g. gene expression levels. May specify quantile in the form of
'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{n_hvgs}{Number of highly variable genes (HVGs) to compute, which will
be used as input to PCA.}

\item{max.iter.harmony}{Maximum number of iterations for Harmony integration.}

\item{seed}{Set a random seed, for reproducibility.}

\item{label}{Whether to label the clusters in 'plot_reduction' space.}

\item{label.size}{Sets size of labels.}

\item{pt.size}{Adjust point size for plotting.}

\item{fig.res}{Figure resolution in ppi (see 'png' function).}

\item{cont_alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max range of values (between 0 and 1).}

\item{discrete_alpha}{Controls opacity of spots. Provide a single alpha value.}

\item{pt.size.factor}{Scale the size of the spots.}

\item{spatial_col_pal}{Continuous colour palette to use from viridis package to
colour spots on tissue, default "inferno".}

\item{crop}{Crop the plot in to focus on points plotted. Set to FALSE to
show entire background image.}

\item{...}{Additional named parameters passed to Seurat's or Harmony
functions.}
}
\value{
An updated Seurat object. Note that if multiple \code{npcs} and \code{ndims}
are given, only the last setting will be returned. All analysis results are
also stored on disk.
}
\description{
This function implements all the analysis steps for perfoming
data integration using Harmony. These include: 1. data processing, e.g.
normalisation, PCA. 2. Running Harmony 3. Perform UMAP and clustering after
data integration. Analysis outputs are stored in corresponding directories.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
