% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPipe_package.R
\docType{package}
\name{SeuratPipe}
\alias{SeuratPipe}
\alias{.datatable.aware}
\title{\code{SeuratPipe}: Streamlining Seurat analysis}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\value{
seurat_pipeline main package documentation.
}
\description{
SeuratPipe contains common analysis, processing and
plotting functions with an attempt to streamline single-cell genomics
analysis using the Seurat package.
}
\author{
C.A.Kapourani \email{kapouranis.andreas@gmail.com}
}
\keyword{datasets}
