% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{subset_feature_plot}
\alias{subset_feature_plot}
\title{Subset feature plot}
\usage{
subset_feature_plot(
  seu,
  subset.by,
  feature,
  reduction = "umap",
  max.cutoff = "q98",
  min.cutoff = NA,
  slot = "data",
  ncol = NULL,
  col_pal = NULL,
  legend.position = "right",
  pt.size = 2,
  pt.stroke = 0.05,
  pt.shape = 21,
  pt.alpha = 1,
  back.pt.size = 0.5,
  back.pt.alpha = 0.1,
  back.pt.color = "grey",
  combine = TRUE,
  dims_plot = c(1, 2),
  order_points_by_value = TRUE,
  ...
)
}
\arguments{
\item{seu}{Seurat object}

\item{subset.by}{Metadata column whose unique values will generate split
feature plots.}

\item{feature}{Feature to plot.}

\item{reduction}{Dimensionality reduction to use.}

\item{max.cutoff}{Maximum cutoff value for feature, may specify quantile
in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{min.cutoff}{Minimum cutoff value for feature, may specify quantile
in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10').}

\item{slot}{Slot to extract data from.}

\item{ncol}{Number of columns for display when having multiple features.}

\item{col_pal}{Continuous colour palette to use, default "RdYlBu".}

\item{legend.position}{Position of legend, default "right" (set to "none"
for clean plot).}

\item{pt.size}{Adjust point size for plotting.}

\item{pt.stroke}{Stroke value for each point.}

\item{pt.shape}{Adjust point shape for plotting.}

\item{pt.alpha}{Adjust alpha value for each point.}

\item{back.pt.size}{Adjust background point size for plotting.}

\item{back.pt.alpha}{Adjust opacity for background points.}

\item{back.pt.color}{Colour for background points.}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}

\item{dims_plot}{Dimensions to plot, must be a two-length numeric vector
specifying x- and y-dimensions.}

\item{order_points_by_value}{Logical, should points be ordered by their value
(e.g. expression levels), which corresponds to plotting on top cells that
have high expression, instead of getting 'buried' by lowly expressed cells.}

\item{...}{Additional parameters passed to ggplot2::geom_point.}
}
\value{
A ggplot2 object.
}
\description{
This function splits the feature plots according to a grouping
variable defined in \code{subset.by}. The main difference is that the whole
dataset is also shown in the background as smaller points.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
